final ISA

+++ stop-remove-container.sh
@@ -5,7 +5,7 @@ container_name=node-sqlite-app-container
 
 > $log_file
 
-if [ "$(docker ps -q -f name=$container_name)" ]; then
+if [ "$(docker ps -a -q -f name=$container_name)" ]; then
     echo "Stopping and removing existing container $container_name..."
     docker stop $container_name >> $log_file 2>&1
     docker rm $container_name >> $log_file 2>&1
